LayoutTests/http/tests/inspector-enabled/dedicated-workers-list-expected.txt

CONSOLE MESSAGE: line 20: Received message from worker 1
CONSOLE MESSAGE: line 20: Received message from worker 2
Tests that dedicated workers created before worker inspection was enabled will be reported to the front-end. Bug 72020

Worker inspection enabled
Added worker: 1
Added worker: 2
Done.
